fuzz: Qualify target names

+++ fuzzit.sh
@@ -1,25 +1,36 @@
 #!/bin/bash
 set -xe
 
-NAME=ulid
-TYPE=$1
+# Validate arguments
+if [ "$#" -ne 1 ]; then
+    echo "Usage: $0 <fuzz-type>"
+    exit 1
+fi
 
-function fuzz {
-    TARGET=$NAME-$1
-    FUNCTION=Fuzz$2
-    go-fuzz-build -libfuzzer -func $FUNCTION -o $NAME.a .
-    clang -fsanitize=fuzzer $NAME.a -o $NAME
-    ./fuzzit create job --type $TYPE $TARGET $NAME
-}
+# Configure
+NAME=ulid
+ROOT=.
+TYPE=$1
 
 # Setup
 export GO111MODULE="off"
 go get -u github.com/dvyukov/go-fuzz/go-fuzz github.com/dvyukov/go-fuzz/go-fuzz-build
 go get -d -v -u ./...
-wget -q -O fuzzit https://github.com/fuzzitdev/fuzzit/releases/download/v2.4.29/fuzzit_Linux_x86_64
-chmod a+x fuzzit
+if [ ! -f fuzzit ]; then
+    wget -q -O fuzzit https://github.com/fuzzitdev/fuzzit/releases/download/v2.4.29/fuzzit_Linux_x86_64
+    chmod a+x fuzzit
+fi
 
 # Fuzz
-fuzz new New
-fuzz parse Parse
-fuzz parse-strict ParseStrict
+function fuzz {
+    FUNC=Fuzz$1
+    TARGET=$2
+    DIR=${3:-$ROOT}
+    go-fuzz-build -libfuzzer -func $FUNC -o fuzzer.a $DIR
+    clang -fsanitize=fuzzer fuzzer.a -o fuzzer
+    ./fuzzit create job --type $TYPE $NAME/$TARGET fuzzer
+}
+fuzz New new
+fuzz NewMonotonic new-monotonic
+fuzz Parse parse
+fuzz ParseStrict parse-strict
